Update the logic of gen2main (#23507)
* Update the logic of gen2main

* Update the logic of gen2main

* Remove useless files

* Rename Az.autorest.hybrid.props to Az.autorest.props

* Rename Az.autorest.hybrid.props to Az.autorest.props

* Rename Az.autorest.hybrid.props to Az.autorest.props

* Rename Az.autorest.hybrid.props to Az.autorest.props

* Move UX folder

* Move UX folder

* Remove ScenarioTest.ResourceManager from az.accounts

+++ tools/PrepareAutorestModule.ps1
@@ -91,39 +91,41 @@ If ($ModuleSet.Contains("Compute"))
 #Region generate the code and make the struture same with main branch.
 $AutorestOutputDir = "$PSScriptRoot\..\artifacts\autorest"
 New-Item -ItemType Directory -Force -Path $AutorestOutputDir
+$FailedModuleList = @()
 foreach ($Module in $ModuleList)
 {
     $RootModuleFolder = "$PSScriptRoot\..\src\$Module\"
-    $ModuleFolders = (Get-ChildItem -path $RootModuleFolder -filter Az.*.psd1 -Recurse).Directory
-    if ($Null -eq $ModuleFolders)
+    $SubModuleFolders = (Get-ChildItem -path $RootModuleFolder -filter Az.*.psd1 -Recurse).Directory
+    if ($Null -eq $SubModuleFolders)
     {
         # Module is not found maybe it's deleted in this PR
         Write-Warning "Cannot find any psd1 files in $RootModuleFolder."
         continue
     }
+    $FailedSubModuleList = @()
     # Go through the module including nested modules.
-    foreach ($ModuleFolder in $ModuleFolders) {
-        Set-Location -Path $ModuleFolder
+    foreach ($SubModuleFolder in $SubModuleFolders) {
+        Set-Location -Path $SubModuleFolder
         # Msbuild will regard autorest's output stream who contains "xx error xx:" as an fault by mistake.
         # We need to redirect output stream to file to avoid the mistake.
-        npx autorest --max-memory-size=8192 >> "$AutorestOutputDir\$Module.log"
+        npx autorest --max-memory-size=8192 >> "$AutorestOutputDir\$Module-SubModuleFolder.log"
         # Exit if generation fails
         if ($lastexitcode -ne 0)
         {
-            exit $lastexitcode
+            $FailedSubModuleList += "$Module-SubModuleFolder"
         }
-
-        ./build-module.ps1
+        else
+        {
+            ./build-module.ps1
+        }
     }
-    $subModuleFolders =  Get-ChildItem -path $RootModuleFolder -Directory -Filter *.Autorest
-    if ($null -eq $subModuleFolders) {
-        write-host "autogen module"
-        Move-Generation2Master -SourcePath "$PSScriptRoot\..\src\$Module\" -DestPath $TmpFolder\src
-    } else {
-        #New-Item -ItemType Directory -Path $TmpFolder\$Module -Force
-        Write-Host "hybrid module"
-        Move-Generation2MasterHybrid -SourcePath "$PSScriptRoot\..\src\$Module\" -DestPath $TmpFolder\src\$Module
+    if ($FailedSubModuleList.Length -ne 0)
+    {
+        $FailedModuleList += $FailedSubModuleList
+        continue
     }
+    #New-Item -ItemType Directory -Path $TmpFolder\$Module -Force
+    Move-Generation2Master -SourcePath "$PSScriptRoot\..\src\$Module\" -DestPath $TmpFolder\src\$Module
     Set-Location -Path $TmpFolder
     Remove-Item "$RootModuleFolder\*" -Recurse -Force
     Copy-Item -Path "$TmpFolder\src\$Module" "$TmpFolder\src\.."  -Recurse -Force
@@ -133,3 +135,13 @@ Remove-Item -Path "$TmpFolder\src" -Recurse -Force
 Remove-Item -Path "$TmpFolder\artifacts" -Recurse -Force
 Remove-Item -Path "$TmpFolder\tools" -Recurse -Force
 Copy-Item "$TmpFolder\*" "$PSScriptRoot\..\src" -Exclude src,.git,tools,build.proj -Recurse -Force
+if ($FailedModuleList.Length -ne 0)
+{
+    Write-Error "Failed to generate code for module(s): $FailedModuleList"
+    foreach ($failedModule in $FailedModuleList)
+    {
+        Write-Error "========= Error log for $failedModule ========="
+        Write-Host (Get-Content "$AutorestOutputDir\$failedModule.log")
+    }
+    exit 1
+}
